backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
pymongo==4.8.0
python-dotenv==1.0.1
PyJWT==2.9.0
Authlib==1.3.1
requests==2.32.3
google-api-python-client==2.139.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.2.0
sendgrid==6.11.0
Jinja2==3.1.6
gunicorn==21.2.0

